Adjusted Output Stdout
stdout apparently doesn't end or finish like a normal write stream, i
assume it stays open for the existence of the process. With that said,
waiting for an end of the lifetime has to happen in two different
places.. or so i figure currently. This sort of sucks, i'd like to
improve this.