NPM and CWD
The cwd needs to be passed in for the process, because npm requires it.
So.. we're doing that. Without it, npm gives odd results because it
looks up the directory chain for a `node_modules` directory and installs
the given directory *into* that directory. Which is like, the opposite
of what we want.