Outdated Prod Code and Tests
There is still an issue that needs to be resolved related to outdated
not returning anything in the given directory if there is no package.

Both install and outdated "suffer" from an issue where if there is
no package.json file, they will walk up the directories until they find
one. Once they find one, they will return the information *from that
package*. This seems incorrect, and counter intuitive to me, so i want
to enforce package.json usage for the given directory.