Improved Outdated Response
This was the first, initial trial for a new response format for outdated
and installed.

npm tends to respond with some *very* weird output, as far as i am
concerned. If outdated or installed is called from the current
directory, and there is no package.json, npm won't fail at all. Instead
it will traverse up the tree looking for a valid project.

In the case of outdated, it will list outdated modules for the way
parent tree.

In the case of install, it will attempt to install the current directory
into the parent tree.

Both of these, in my opinion, are crazy and horribly unintended actions.
As such, the interface i am choosing is to fail if the given directory
is not a valid project.