For a brief moment FS and ChildProcess Rules
Made a handful of improvements to the code, and some additions to
the abstractions.

I *was* going to remove the transports abstraction, and instead use fs
and exec. Then i remembered that the target installations won't have
access to `kdc outdated`, so these fs/exec changes will be removed in
the soonafter commits.