Added Option for Github Only Feature
As of the time of this commit, the `mode:true` option is not supported
in the official repo, and only exists in the master branch of the
grunt-contrib-copy project on github. This feature is not likely to be
required, it's just good practice, so i am not modifying our package
requirements and rather simply adding the option to true.

If the host platform has it installed the option gets used, otherwise it
gets ignored.