Updated the binGen and Commented a Test
The binGen had some faulty logic in it, so i fixed that.

Also, there is a bit of an annoyance with Commander.js in that i am
unsure how to easily calculate invalid commands. This is likely related
to a recent bug discovery, so until the whole thing is better handled in
Commander.js i am just going to remove this "feature" all together, and
thus comment out the test for it as well.