Bare, and Closure
Made the coffee compiles bare by default, although i think this is a
poor design. For now, we are doing this for compatibility with kdc.

In the long run it might be nice to add an app specific object into
the main closure for each app.. though with automatic ghosting from
coffee-script this may be challenging.

Either way, we're compatible for now, so there's that.