Even Better Bare and Closure
To settle my dislike for defaulting to bare for all coffee compiles,
bare is now added by default only to the legacy `kdc` compile. With
this comes an added --bare option which will be passed to
Coffee.

The compiled version comment has also been fixed, and works correctly.

The final closure now comes with a semi-colon, because.. why not?

And finally, the end event conflict between stdout and fsWritable
has been resolved and now both methods are handled properly, friendly,
and .. happily. or something.