Bare Coffee
Normally the coffee closures don't really matter on Node one way or the
other, but the recent addition of rewire will cause the compiled tests
to fail if they are closured. Not exactly sure on the specific cause,
but it obviously has to do with how rewire mocks the test objects.

So, we're enabling coffee-bare to ensure our mocks work when compiled.