Added File Opts
I'm not happy with the state of this addition, due to the nature of the
"loaders". The issue is that commonjs loader requires transforms, which
means i can't just load and pipe the file. So, in a very inefficient way
i am making every file into a single loader for now, and applying all
optional compiles (coffee/etc) onto it.

This is slow and rather ugly, but based on the Commonjs restriction it's
what we're going with for now.